+++ block-production-albatross/tests/mod.rs
@@ -216,7 +216,7 @@ fn it_can_produce_macro_blocks() {
 
     let block = sign_macro_block(&keypair, proposal, Some(extrinsics));
     assert_eq!(
-        blockchain.push_block(Block::Macro(block)),
+        blockchain.push(Block::Macro(block)),
         Ok(PushResult::Extended)
     );
 }
@@ -243,7 +243,7 @@ fn it_can_produce_election_blocks() {
 
         let block = sign_macro_block(&keypair, proposal, Some(extrinsics));
         assert_eq!(
-            blockchain.push_block(Block::Macro(block)),
+            blockchain.push(Block::Macro(block)),
             Ok(PushResult::Extended)
         );
     }
